requirements.txt
Django==2.0
django-allauth==0.23.0
django-bootstrap3==6.2.0
django-rest-auth==0.6.0
django-rest-framework-docs==0.1.7
django-rest-swagger==0.3.3
djangorestframework==3.12.1
requests==2.7.0